+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 **?.?.? Changes:**
 
+* Combo:
+  * Fixed effect selection ignoring incompatibility rules
+
 * Renamed effect: Touch Void -> Voidtouch Everyone
 
 **1.8.0 Changes:**